refactor code as discussed offline
